# docs/swagger.yaml
definitions:
...
  json.CategoriaLojistaResponse:
    properties:
      id:
        type: integer
      nome:
        type: string
    type: object
...
  json.CategoriasLojistaResponse:
    properties:
      categorias:
        items:
          $ref: '#/definitions/json.CategoriaLojistaResponse'
        type: array
      total:
        type: integer
    type: object
...
  json.LojaRequest:
    properties:
      cnpj:
        type: string
      id_categoria:
        type: integer
      imagem:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      nome:
        type: string
    required:
    - cnpj
    - id_categoria
    - latitude
    - longitude
    - nome
    type: object
  json.LojaResponse:
    properties:
      categoria:
        type: string
      cnpj:
        type: string
      distancia:
        description: Distância em km
        type: number
      id:
        type: integer
      id_categoria:
        type: integer
      imagem:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      nome:
        type: string
    type: object
...
  json.LojasResponse:
    properties:
      lojas:
        items:
          $ref: '#/definitions/json.LojaResponse'
        type: array
      total:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /lojas:
    get:
      consumes:
      - application/json
      description: Retorna uma lista com todas as lojas ativas
      produces:
      - application/json
      responses:
        "200":
          description: Lista de lojas
          schema:
            items:
              $ref: '#/definitions/json.LojaResponse'
            type: array
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Lista todas as lojas
      tags:
      - Lojas
    post:
      consumes:
      - application/json
      description: Cria uma nova loja com todos os dados fornecidos
      parameters:
      - description: Dados completos da loja
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/json.LojaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Loja criada com sucesso
          schema:
            $ref: '#/definitions/json.LojaResponse'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Criação da loja completa
      tags:
      - Lojas
  /lojas/{id}:
    delete:
      consumes:
      - application/json
      description: Realiza soft delete da loja, marcando como excluída sem remover
        do banco
      parameters:
      - description: ID da loja
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Loja excluída com sucesso
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Loja não encontrada
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Exclui loja (soft delete)
      tags:
      - Lojas
    get:
      consumes:
      - application/json
      description: Retorna os dados de uma loja específica pelo ID
      parameters:
      - description: ID da loja
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Loja encontrada
          schema:
            $ref: '#/definitions/json.LojaResponse'
        "400":
          description: ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Loja não encontrada
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Busca loja por ID
      tags:
      - Lojas
    put:
      consumes:
      - application/json
      description: Atualiza os dados de uma loja existente
      parameters:
      - description: ID da loja
        in: path
        name: id
        required: true
        type: integer
      - description: Dados atualizados da loja
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/json.LojaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Loja atualizada com sucesso
          schema:
            $ref: '#/definitions/json.LojaResponse'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Loja não encontrada
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Atualiza loja
      tags:
      - Lojas
...
  /lojas/{id}/restore:
    post:
      consumes:
      - application/json
      description: Restaura uma loja que foi soft deleted
      parameters:
      - description: ID da loja
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Loja restaurada com sucesso
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Loja não encontrada
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Restaura loja excluída
      tags:
      - Lojas
...
  /lojas/categorias:
    get:
      consumes:
      - application/json
      description: Retorna todas as categorias de lojista disponíveis
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/json.CategoriasLojistaResponse'
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Lista categorias de lojista
      tags:
      - Lojas
  /lojas/proximidade:
    get:
      consumes:
      - application/json
      description: Retorna lista de lojas ordenadas por proximidade do usuário
      parameters:
      - description: Latitude do usuário
        in: query
        name: latitude
        required: true
        type: number
      - description: Longitude do usuário
        in: query
        name: longitude
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/json.LojasResponse'
        "400":
          description: Parâmetros inválidos
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Lista lojas por proximidade
      tags:
      - Lojas
...
swagger: "2.0"